Organize linalg rewrites according to library
