+++ pytensor/tensor/rewriting/linalg/utils.py
@@ -1,0 +1,84 @@
+import logging
+
+from pytensor import tensor as pt
+from pytensor.graph import Constant
+from pytensor.graph.rewriting.basic import node_rewriter
+from pytensor.graph.rewriting.unify import OpPattern
+from pytensor.scalar.basic import Mul
+from pytensor.tensor._linalg.inverse import MatrixInverse, MatrixPinv
+from pytensor.tensor.basic import (
+    Eye,
+    TensorVariable,
+    atleast_Nd,
+    diagonal,
+)
+from pytensor.tensor.elemwise import DimShuffle, Elemwise
+from pytensor.tensor.math import variadic_mul
+from pytensor.tensor.rewriting.basic import (
+    register_canonicalize,
+    register_specialize,
+    register_stabilize,
+)
+
+
+logger = logging.getLogger(__name__)
+
+MATRIX_INVERSE_OPS = (MatrixInverse, MatrixPinv)
+
+
+def matrix_diagonal_product(x):
+    return pt.prod(diagonal(x, axis1=-2, axis2=-1), axis=-1)
+
+
+def is_matrix_transpose(x: TensorVariable) -> bool:
+    """Check if a variable corresponds to a transpose of the last two axes"""
+    match x.owner_op:
+        case DimShuffle(is_left_expanded_matrix_transpose=True):
+            return True
+    return False
+
+
+def is_eye_mul(x) -> None | tuple[TensorVariable, TensorVariable]:
+    # Check if we have a Multiplication with an Eye inside
+    # Note: This matches for cases like (eye * 0)!
+    match x.owner_op_and_inputs:
+        case Elemwise(Mul()), *mul_inputs:
+            pass
+        case _:
+            return None
+
+    x_bcast = x.type.broadcastable[-2:]
+    eye_input = None
+    non_eye_inputs = []
+    for mul_input in mul_inputs:
+        # We only care about Eye if it's not broadcasting in the multiplication
+        if mul_input.type.broadcastable[-2:] == x_bcast:
+            match mul_input.owner_op_and_inputs:
+                case (Eye(), _, _, Constant(0)):
+                    eye_input = mul_input
+                    continue
+                # This whole condition checks if there is an Eye hiding inside a DimShuffle.
+                # This arises from batched elementwise multiplication between a tensor and an eye, e.g.:
+                # tensor(shape=(None, 3, 3) * eye(3). This is still potentially valid for diag rewrites.
+                case (DimShuffle(is_left_expand_dims=True), ds_input):
+                    match ds_input.owner_op_and_inputs:
+                        case (Eye(), _, _, Constant(0)):
+                            eye_input = mul_input
+                            continue
+        # If no match:
+        non_eye_inputs.append(mul_input)
+
+    if eye_input is None:
+        return None
+
+    return eye_input, variadic_mul(*non_eye_inputs)
+
+
+@register_canonicalize
+@register_stabilize
+@register_specialize
+@node_rewriter([OpPattern(DimShuffle, is_left_expanded_matrix_transpose=True)])
+def useless_symmetric_transpose(fgraph, node):
+    x = node.inputs[0]
+    if getattr(x.tag, "symmetric", False):
+        return [atleast_Nd(x, n=node.outputs[0].type.ndim)]
